Cleaning up the store.
