+++ resources/assets/init.js
@@ -15,7 +15,7 @@ import React from 'react';
 import ReactDom from 'react-dom';
 import thunk from 'redux-thunk';
 import { routerReducer, routerMiddleware } from 'react-router-redux';
-import { configureStore } from './store';
+import { configureStore } from './store/store';
 import * as reducers from './reducers';
 
 // Browser polyfills
