+++ resources/assets/middleware/api.js
@@ -1,0 +1,7 @@
+const apiMiddleware = ({ getState, dispatch }) => next => action => {
+  console.log(action.type);
+  next(action);
+  console.log(action.type);
+};
+
+export default apiMiddleware;
