+++ resources/assets/store/middlewares.js
@@ -1,0 +1,9 @@
+import experimentsMiddleware from '../middleware/experiments';
+import experimentsApiMiddleware from '../middleware/experimentsApi';
+
+const middlwares = [
+  experimentsApiMiddleware,
+  experimentsMiddleware,
+];
+
+export default middlwares;
